requirements.txt
# Core dependencies
python-telegram-bot[callback-data]>=20.7
redis>=5.0.1
notion-client>=2.2.1
typesense>=0.17.0
pydantic>=2.5.1
pydantic-settings>=2.1.0

# Async support
aiohttp>=3.9.1
asyncio>=3.4.3

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
httpx>=0.25.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0

# Development
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1

# Type stubs
types-redis>=4.6.0.20231227
types-pytz>=2023.3.1.1 